requirements.txt
APScheduler==3.10.4
av==12.3.0
Fabric==3.2.2
Flask==3.0.3
groq==0.11.0
numpy==2.1.1
pexpect==4.9.0
phidata==2.4.27
Pillow==10.4.0
playwright==1.45.1
pytesseract==0.3.13
streamlit==1.38.0
streamlit_webrtc==0.47.9
typer==0.12.5
